Allow eshell-toggle to use Git root directory.
This commit enables `eshell-toggle--get-directory` to fetch root
directory if inside Git project. The user can turn this option on by
setting `eshell-toggle-use-git-root` different from `nil`.

This option is `nil` by default.

In the case when both `eshell-toggle-use-projectile-root` and
`eshell-toggle-use-git-root` are enabled and inside Git project the
behaviour should be to default to Projectile root directory.